# apps/user-service/swagger.json
{"openapi":"3.0.0","paths":{"/api/users/{userId}":{"patch":{"operationId":"createOrUpdateUserAsync","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequestDto"}}}},"responses":{"200":{"description":""}},"tags":["User"]},"post":{"operationId":"updateUserAsync","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequestDto"}}}},"responses":{"201":{"description":""}},"tags":["User"]},"get":{"operationId":"getUserAsync","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"]}},"/api/users":{"get":{"operationId":"getAllUsersAsync","parameters":[],"responses":{"200":{"description":""}},"tags":["User"]}},"/api/users/userId":{"delete":{"operationId":"deleteUserAsync","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"]}},"/api/users/{userId}/data-sources/{dataSourceId}":{"patch":{"operationId":"createOrUpdateDataSourceUnderUser","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}},{"name":"dataSourceId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataSource"}}}},"responses":{"200":{"description":""}},"tags":["DataSource"]},"post":{"operationId":"updateDataSourceUnderUserAsync","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}},{"name":"dataSourceId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataSource"}}}},"responses":{"201":{"description":""}},"tags":["DataSource"]},"get":{"operationId":"getDataSourceUnderUserAsync","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}},{"name":"dataSourceId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["DataSource"]},"delete":{"operationId":"deleteDataSourceUnderUserAsync","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}},{"name":"dataSourceId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["DataSource"]}},"/api/users/{userId}/data-sources":{"get":{"operationId":"getDataSourcesUnderUserAsync","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["DataSource"]}}},"info":{"title":"user-service","description":"","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"UserRequestDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"emailId":{"type":"string"},"password":{"type":"string"}},"required":["firstName","lastName","emailId","password"]},"DataSource":{"type":"object","properties":{"name":{"type":"string","description":"Data source name"},"host":{"type":"string","description":"Data source URL host"},"resourcePath":{"description":"Path array to the resource","type":"array","items":{"type":"string"}},"queryParameters":{"description":"Query parameters array","type":"array","items":{"type":"string"}},"queryValues":{"description":"Values for the query parameters","type":"array","items":{"type":"string"}},"accessToken":{"type":"string","description":"Access token for the API"},"refreshTokenResourcePath":{"description":"Resource path for access token refresh","type":"array","items":{"type":"string"}},"refreshTokenQueryParameters":{"description":"Query parameters for refresh token API","type":"array","items":{"type":"string"}},"refreshTokenQueryValues":{"description":"Values for refresh token query parameters","type":"array","items":{"type":"string"}}},"required":["name","host","resourcePath","queryParameters","queryValues","accessToken","refreshTokenResourcePath","refreshTokenQueryParameters","refreshTokenQueryValues"]}}}}